multiplayer refactoring + added client only helper
