+++ src/Plugin.cs
@@ -5,6 +5,7 @@ using BepInEx.Configuration;
 using BepInEx.Logging;
 using PolyMod.Android;
 using PolyMod.Managers;
+using PolyMod.Multiplayer;
 using UnityEngine;
 
 namespace PolyMod;
@@ -136,6 +137,7 @@ public partial class Plugin : BepInEx.Unity.IL2CPP.BasePlugin
 
 		Main.Init();
 		Multiplayer.Client.Init();
+		ModMultiplayer.Init();
 		AndroidHandler.Init();
 	}
